优化目录

+++ src/LinFx/Application/Abstractions/IHasTotalCount.cs
@@ -8,6 +8,6 @@
         /// <summary>
         /// Total count of Items.
         /// </summary>
-        int TotalCount { get; set; }
+        long TotalCount { get; set; }
     }
 }
